shield_exceptions should never have been in common in the first place, but it's only called in one place now so we don't need a separate target for it
