+++ browser/net/brave_site_hacks_network_delegate_helper.cc
@@ -14,7 +14,6 @@
 #include "base/no_destructor.h"
 #include "base/strings/string_util.h"
 #include "brave/common/network_constants.h"
-#include "brave/common/shield_exceptions.h"
 #include "brave/common/url_constants.h"
 #include "brave/components/brave_shields/browser/brave_shields_util.h"
 #include "content/public/common/referrer.h"
@@ -28,6 +27,18 @@
 namespace brave {
 
 namespace {
+
+bool IsUAWhitelisted(const GURL& gurl) {
+  static std::vector<URLPattern> whitelist_patterns({
+    URLPattern(URLPattern::SCHEME_ALL, "https://*.duckduckgo.com/*"),
+    // For Widevine
+    URLPattern(URLPattern::SCHEME_ALL, "https://*.netflix.com/*")
+  });
+  return std::any_of(whitelist_patterns.begin(), whitelist_patterns.end(),
+      [&gurl](URLPattern pattern){
+        return pattern.MatchesURL(gurl);
+      });
+}
 
 const std::string& GetQueryStringTrackers() {
   static const base::NoDestructor<std::string> trackers(base::JoinString(
