--- common/shield_exceptions.cc
@@ -1,28 +1,0 @@
-/* Copyright (c) 2019 The Brave Authors. All rights reserved.
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this file,
- * You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-#include "brave/common/shield_exceptions.h"
-
-#include <algorithm>
-#include <vector>
-
-#include "extensions/common/url_pattern.h"
-#include "url/gurl.h"
-
-namespace brave {
-
-bool IsUAWhitelisted(const GURL& gurl) {
-  static std::vector<URLPattern> whitelist_patterns({
-    URLPattern(URLPattern::SCHEME_ALL, "https://*.duckduckgo.com/*"),
-    // For Widevine
-    URLPattern(URLPattern::SCHEME_ALL, "https://*.netflix.com/*")
-  });
-  return std::any_of(whitelist_patterns.begin(), whitelist_patterns.end(),
-      [&gurl](URLPattern pattern){
-        return pattern.MatchesURL(gurl);
-      });
-}
-
-}  // namespace brave
